chg: improve paper.
