+++ docs/paper/Makefile
