+++ docs/paper/images/annotations.png
