+++ docs/paper/images/inscriptis-vs-lynx.png
