+++ docs/paper/images/inscriptis-vs-lynx.xcf
