+++ docs/paper/images/raw/inscriptis.png
