+++ docs/paper/images/raw/lynx.png
