+++ src/inscriptis/metadata.py
@@ -1,0 +1,7 @@
+"""Inscriptis metadata information."""
+
+__author__ = 'Albert Weichselbraun, Fabian Odoni'
+__author_email__ = 'albert.weichselbraun@fhgr.ch, fabian.odoni@fhgr.ch'
+__copyright__ = '2016-2021 Albert Weichselbraun, Fabian Odoni'
+__license__ = 'Apache 2.0'
+__version__ = '2.0rc1'
