+++ tests/test_broken_table_handling.py
@@ -5,7 +5,7 @@
 Tests the handling of tables that do not properly close all column tags.
 """
 
-from inscriptis.engine import get_text
+from inscriptis import get_text
 from inscriptis.css_profiles import CSS_PROFILES
 from inscriptis.model.config import ParserConfig
