+++ tests/test_html_snippets.py
@@ -7,7 +7,7 @@ contains the reference conversion.
 from os.path import dirname, join
 from glob import glob
 
-from inscriptis.engine import get_text
+from inscriptis import get_text
 from inscriptis.css_profiles import CSS_PROFILES
 from inscriptis.model.config import ParserConfig
