+++ tests/test_list_div.py
@@ -6,7 +6,7 @@
     or space between the tags.
 """
 
-from inscriptis.engine import get_text
+from inscriptis import get_text
 from inscriptis.css_profiles import CSS_PROFILES
 from inscriptis.model.config import ParserConfig
