+++ tests/test_margin_before_at_start.py
@@ -5,7 +5,7 @@
     or space between the tags.
 """
 
-from inscriptis.engine import get_text
+from inscriptis import get_text
 
 
 def test_content():
