+++ tests/test_white_space_handling.py
@@ -5,7 +5,7 @@
 Tests different white-space handling.
 """
 
-from inscriptis.engine import get_text
+from inscriptis import get_text
 from inscriptis.css_profiles import CSS_PROFILES
 from inscriptis.model.config import ParserConfig
